# Use accurate benchmark for Koffi